LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-serialization/cssUnparsedValue-expected.txt


FAIL CSSUnparsedValue containing strings serializes to its concatenated contents assert_equals: expected "lemonade" but got "lem,on,ade"
FAIL CSSUnparsedValue containing variable references serializes its concatenated contents Can't find variable: CSSVariableReferenceValue
FAIL CSSUnparsedValue containing mix of strings and variable references serializes to its concatenated contents Can't find variable: CSSVariableReferenceValue
